編集ページ表示データが入ってない

+++ src/main/java/com/example/repository/EventRepository.java
@@ -2,14 +2,16 @@ package com.example.repository;
 
 import java.util.List;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import com.example.entity.Event;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
-
+public interface EventRepository extends JpaRepository<Event, Integer> {
+	
+	
 
     @Query("SELECT e FROM Event e WHERE YEAR(e.startevent) = :year AND MONTH(e.startevent) = :month")
     List<Event> findEventsForMonth(@Param("year") int year, @Param("month") int month);
